requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Supabase integration
supabase==2.0.0
postgrest==0.11.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Audio/Video processing
ffmpeg-python==0.2.0
librosa==0.10.1
pyannote.audio==3.1.1
soundfile==0.12.1

# Machine Learning
openai-whisper==20231117
tensorflow==2.15.0
torch==2.1.1
transformers==4.36.0

# LangChain ecosystem
langchain==0.1.0
langgraph==0.0.26
langchain-openai==0.0.2
langchain-google-genai==0.0.6

# HTTP and networking
httpx==0.25.2
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.1
flake8==6.1.0

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.8

# Environment management
python-dotenv==1.0.0

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Async support
asyncpg==0.29.0 